+++ src/bstone/src/bstone_audio_mixer.h
@@ -89,7 +89,7 @@ struct AudioMixerInitParam
 {
 	AudioDriverType audio_driver_type;
 
-	OplEmulatorType opl3_type;
+	OplEmulatorType opl_emulator_type;
 	int dst_rate;
 
 	int mix_size_ms; // (milliseconds)
@@ -114,7 +114,7 @@ public:
 	AudioMixer() = default;
 	virtual ~AudioMixer() = default;
 
-	virtual OplEmulatorType get_opl3_type() const = 0;
+	virtual OplEmulatorType get_opl_emulator_type() const = 0;
 	virtual int get_rate() const = 0;
 	virtual int get_channel_count() const = 0;
 	virtual int get_mix_size_ms() const = 0;