+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [SDL] Raise minimum version to v3.4.
 - Image extractor writes images in PNG file format.
 - Rename value 'adlib' to 'opl' of CVar 'snd_sfx_type'.
+- Rename CVar 'snd_opl3_type' to 'snd_opl_emulator'.
 
 ### Removed
 - [CMake] Variable SDL2W_SDL2_DIR.